+++ src/main/java/kr/codesqaud/cafe/exception/user/MismatchedPasswordException.java
@@ -1,4 +1,4 @@
-package kr.codesqaud.cafe.exception;
+package kr.codesqaud.cafe.exception.user;
 
 public class MismatchedPasswordException extends RuntimeException {
     public MismatchedPasswordException() {
